BatteryService: Add support for oem fast charger detection
Based on AOSPA fw/base 0af3194
  - AOSPA/android_frameworks_base@0af3194

Allows to indicate, whether a device is charged using a
proprietary oem fast charge solution, which write their status
to sysfs.

The oem fast charge detection tries to be as generic as possible and is
configured via overlays.

Paths to sysfs to read status
  - core/res/res/values/lineage_config.xml
    - config_oemFastChargerStatusPaths